Move schema-block utility into packages directory
