+++ app/models/schema-block.ts
@@ -1,31 +1,6 @@
 import { attr } from '@ember-decorators/data';
-
+import { SchemaBlock, SchemaBlockType } from 'ember-osf-web/packages/registration-schema/schema-block';
 import OsfModel from './osf-model';
-
-type SchemaBlockType =
-    'page-heading' |
-    'section-heading' |
-    'subsection-heading' |
-    'paragraph' |
-    'question-label' |
-    'short-text-input' |
-    'long-text-input' |
-    'file-input' |
-    'contributors-input' |
-    'multi-select-input' |
-    'single-select-input' |
-    'select-input-option' |
-    'select-other-option';
-
-export interface SchemaBlock {
-    blockType?: SchemaBlockType;
-    schemaBlockGroupKey?: string;
-    registrationResponseKey?: string;
-    displayText?: string;
-    helpText?: string;
-    required?: boolean;
-    index?: number;
-}
 
 export default class SchemaBlockModel extends OsfModel implements SchemaBlock {
     @attr('string') blockType?: SchemaBlockType;
