+++ app/packages/registration-schema/get-schema-block-group.ts
@@ -1,41 +1,12 @@
 import { assert } from '@ember/debug';
-import { SchemaBlock } from 'ember-osf-web/models/schema-block';
-
-export interface SchemaBlockGroup {
-  labelBlock?: SchemaBlock;
-  inputBlock?: SchemaBlock;
-  optionBlocks?: SchemaBlock[];
-  schemaBlockGroupKey?: string;
-  registrationResponseKey?: string;
-}
+import { SchemaBlock } from 'ember-osf-web/packages/registration-schema/schema-block';
+import { SchemaBlockGroup } from 'ember-osf-web/packages/registration-schema/schema-block-group';
 
 function isEmpty(input: string | undefined) {
     if (input === undefined || input === null || input === '') {
         return true;
     }
     return false;
-}
-
-export function getPages(blocks: SchemaBlock[]) {
-    const pageArray = blocks.reduce(
-        (pages, block) => {
-            // instantiate first page if the schema doesn't start with a page-heading
-            if (pages.length === 0 && block.blockType !== 'page-heading') {
-                const blankPage: SchemaBlock[] = [];
-                pages.push(blankPage);
-            }
-
-            const lastPage: SchemaBlock[] = pages.slice(-1)[0];
-            if (block.blockType === 'page-heading') {
-                pages.push([block]);
-            } else {
-                lastPage.push(block);
-            }
-            return pages;
-        },
-        [] as SchemaBlock[][],
-    );
-    return pageArray;
 }
 
 export function getSchemaBlockGroup(blocks: SchemaBlock[], key: string) {
