+++ app/packages/registration-schema/schema-block-group.ts
@@ -1,0 +1,9 @@
+import { SchemaBlock } from 'ember-osf-web/packages/registration-schema/schema-block';
+
+export interface SchemaBlockGroup {
+    labelBlock?: SchemaBlock;
+    inputBlock?: SchemaBlock;
+    optionBlocks?: SchemaBlock[];
+    schemaBlockGroupKey?: string;
+    registrationResponseKey?: string;
+  }
