+++ tsconfig.json
@@ -88,6 +88,7 @@
     "config",
     "tests",
     "mirage",
+    "app/packages",
     "types",
     "lib/analytics-page",
     "lib/collections",
